% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{as.trajectory}
\alias{as.trajectory}
\title{Convert to Trajectory Object}
\usage{
as.trajectory(x, delta_time = 1)
}
\arguments{
\item{x}{An object of class `data.frame` with `x` and `y` columns.}

\item{delta_time}{Control variable with a default value of 1.}
}
\value{
A `Trajectory` object containing the coordinates from the `data.frame`.
}
\description{
Converts an object of type `data.frame` or similar into a `Trajectory` object. The `data.frame`
should contain columns `x` and `y` representing the coordinates.
}
