% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Calculate the Centroid of Object}
\usage{
centroid(obj)
}
\arguments{
\item{obj}{An object of class `trajectory` that must contain a component named `points`.}
}
\value{
A numeric vector containing the coordinates of the centroid of the points stored in `obj$points`.
}
\description{
This function calculates the centroid of a set of points contained in an object of class 
`trajectory`. The centroid is the average of the coordinates of all the points.
}
