% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceTraveled.R
\name{distanceTraveled}
\alias{distanceTraveled}
\title{Calculates the total distance traveled through a series of points.}
\usage{
distanceTraveled(data)
}
\arguments{
\item{data}{An object of class `trajectory` containing a collection of points with coordinates.}
}
\value{
A numeric value representing the total distance traveled through the points.
}
\description{
Computes the cumulative distance between consecutive points in a given set of coordinates. It sums the Euclidean distances between each pair of adjacent 
points, providing the total distance traveled along the path defined by the points.
}
