% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapFromTrajectory.R
\name{heatmapFromTrajectory}
\alias{heatmapFromTrajectory}
\title{Creates a heatmap from a trajectory.}
\usage{
heatmapFromTrajectory(data, plim = NULL, plot.pal = TRUE, ...)
}
\arguments{
\item{data}{An object of class `trajectory` containing a collection of points with coordinates.}

\item{plim}{Optional vector of length two specifying the plotting limits for the density 
values. If provided, values outside this range will be adjusted.}

\item{plot.pal}{A logical value indicating whether to plot the color palette alongside 
the heatmap (default is TRUE).}

\item{...}{Additional graphical parameters to customize the image.}
}
\value{
A matrix of density values corresponding to the heatmap.
}
\description{
Generates a heatmap based on the density of points in a trajectory. 
It utilizes kernel density estimation to visualize the concentration of points, 
allowing for adjustments in color palette and plotting limits.
}
