% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateTrajectory.R
\name{interpolateTrajectory}
\alias{interpolateTrajectory}
\title{Interpolates a trajectory}
\usage{
interpolateTrajectory(data, n = 4)
}
\arguments{
\item{data}{An object of class `trajectory` containing a collection of points with coordinates.}

\item{n}{An integer indicating the number of intermediate points to generate
between each pair of points. Default is 4.}
}
\value{
An object of class `trayectoria`, which contains the `x` and `y` coordinates, including
        both the original points and any interpolated points. The `trayectoria` object allows users to 
        store, manipulate, and analyze the coordinates of the object across frames. If no object is detected 
        in a frame, the corresponding coordinates will be set to `NA`.
}
\description{
Given a dataset of points, it interpolates to generate intermediate points between them.
}
