% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{is.trajectory}
\alias{is.trajectory}
\title{Check if Object is a Trajectory}
\usage{
is.trajectory(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
A logical value (`TRUE` or `FALSE`) indicating whether the object is of class `Trajectory`.
}
\description{
Checks if an object is of class `Trajectory`.
}
