% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{length.trajectory}
\alias{length.trajectory}
\title{Length of the Trajectory}
\usage{
\method{length}{trajectory}(x)
}
\arguments{
\item{x}{An object of class `Trajectory`.}
}
\value{
A numeric value representing the total length of the trajectory.
}
\description{
Calculates the total length of the trajectory, which is the sum of the distances between 
consecutive points in the `Trajectory` object.
}
