% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalVectors.R
\name{normalVectors}
\alias{normalVectors}
\title{Calculate the normals from tangents of a trajectory object}
\usage{
normalVectors(obj)
}
\arguments{
\item{obj}{An object of class `trajectory` that contains a component `points`.}
}
\value{
A matrix of the same dimensions as `obj$points` with the normal vectors. 
Each row represents a normal vector at a given point, which is obtained by rotating 
the corresponding tangent vector by 90 degrees.
}
\description{
This function computes the normal vectors at each point of an object of class 
`trajectory` by rotating the tangent vectors 90 degrees.
}
