% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{or}
\alias{or}
\alias{|.trajectory}
\title{Combine Two Trajectories}
\usage{
\method{|}{trajectory}(tr1, tr2)
}
\arguments{
\item{tr1}{A `Trajectory` object.}

\item{tr2}{A `Trajectory` object.}
}
\value{
A new `Trajectory` object containing the concatenated points of the two input trajectories.
}
\description{
Combines two `Trajectory` objects by concatenating their points. The result is a new trajectory 
that contains the coordinates of both input trajectories.
}
