% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot.trajectory}
\alias{plot.trajectory}
\title{Plot the Trajectory}
\usage{
\method{plot}{trajectory}(x, ..., stepSize = 0)
}
\arguments{
\item{x}{A `Trajectory` object to plot.}

\item{...}{Other parameters to be passed.}

\item{stepSize}{An integer specifying the interval for plotting segments of the trajectory. 
If set to 0, the entire trajectory is plotted. Default is 0.}
}
\value{
A plot displaying the trajectory.
}
\description{
Plots the trajectory by displaying the `x` and `y` coordinates as points on a 2D plot. Optionally, 
you can add a line connecting the points.
}
