% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{print.summary.trajectory}
\alias{print.summary.trajectory}
\title{Print Summary of the Trajectory}
\usage{
\method{print}{summary.trajectory}(x, ...)
}
\arguments{
\item{x}{A `summary.trajectory` object, typically returned by the `summary.trajectory` method.}

\item{...}{Other parameters to be passed.}
}
\value{
Prints the summary directly.
}
\description{
Prints a detailed summary of the trajectory, including statistics for both the `x` and `y` 
coordinates and the total number of frames.
}
