% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{summary.trajectory}
\alias{summary.trajectory}
\title{Summary of the Trajectory}
\usage{
\method{summary}{trajectory}(object, ...)
}
\arguments{
\item{object}{An object of class `Trajectory`.}

\item{...}{Other parameters to be passed.}
}
\value{
A summary object containing statistics for both the `x` and `y` coordinates of the 
        trajectory.
}
\description{
Generates a summary of the trajectory, including the minimum, maximum, and mean of the `x` 
and `y` coordinates, as well as the total number of frames.
}
