% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangentVectors.R
\name{tangentVectors}
\alias{tangentVectors}
\title{Calculate the tangent cectors of a trajectory object}
\usage{
tangentVectors(obj)
}
\arguments{
\item{obj}{An object of class `trajectory` that contains a component `points`.}
}
\value{
A matrix where each row is a unit tangent vector corresponding to each pair of 
consecutive points in `obj$points`.
}
\description{
This function calculates the unit tangent vectors at each point of the trajectory represented 
by a `trajectory` object. The tangent vectors are computed by taking the difference between 
consecutive points and normalizing them to have unit length.
}
